Fix native library name for Linux and macOS (#39)
* Fix native library name for Linux and macOS

Building on Linux names the file "libstructs.so". I haven't tested on macOS, but judging from the other ffi samples it seems it should also be renamed with the "lib" prefix.

* Fix native library name in structs_test.dart

In ead90d96bf348a19dcf1015218133c36a96b85b6 I forgot to update the test file to also include the "lib" prefix for the native library name.